feature: ISyncTarget provides supported capabilities
Prior to start the synchronisation process, the capabilities
of the SyncTarget is checked, whether its support for multiple
languages complies with the list of languages to be synced.
In case multiple languages are set to be synced but the SyncTarget
does not support multiple languages, an exception is thrown and the
synchronization job is aborted.

The general availability of configured languages subject to sync
is checked with the import settings model `SettingsDatabaseModel`
and its function `ValidateForSync()`

see: https://trello.com/c/J5yPhFV9/735-add-setting-if-multi-language-is-supported-throw-exception-if-the-restriction-is-violated