feature: introduce basic DocFX infrastructure for documentation
In order to create a web site for documentation from the
tripe-slash comments within the source code, the tool
[`DocFX`](dotnet.github.io/docfx/) is used.

In order to drive the automatic creation of the documentation pages,
additional custom tasks are added to the build configuration file.

The documentation pages are split into two sections. The first is
for general information (called "general") and will contain
how-to, example, etc.

The second parts covers the API documentation. Since the created web pages
need to cover various versions of APIs, the current major version of
the API is added to the URL path. Then API documentation for the old
majore release can be maintained and kept available, while development
of next major release is continuing.