fix: SettingsDatabaseModel check ChannelId to be greater than 0
The `ChannelId` must be greater than `0` as channels use a positive
number. So, values equal or below `0` can not be valid channels.
Hence a test on these value has been added to check for such invalid
values.

Yet it does not test, whether the channel ID itself is valid. Only
the syntax is checked. Positive channel IDs might still be invalid
because the channel does not exist (anymore).