feature: (test) introduce test coverage report
Test coverage information files `lcov.info` and
`lcov.cobertura.xml` are created at every test-run.

The additional target `Coverage` creates a local
coverage report. In order to crate the local report
you need to call the custom target like this:

```
    dotnet msbuild -t:Coverage
```

Only the unit test files are covered with this report.
Integration tests are not included in the report
at the moment.