feature: add English fallback for metadata
If English is not requested as an import language then
it is missing as a fallback. In that case all languages
with missing values are filled by using the available
English version.

However, if English is part of the requested import languages,
then the sync target is responsible for taking care of
displaying English translations as fallback. Then no
missing language version is filled with the English version.