fix: make CLAPI client use refreshed access token
In case access to Smint.io API failed, because access token
had expired, a new one was fetched. Unfortunately the API
client has never been updated with the new access token.
So synchronizing failed for ever.

This might be the reason why a new OAuth authorization
is performed all the time starting the sync application.
Because of that, this error has not been detected so far.

Not it is fixed ...