feature: (BREAKING) use downloader instead of passing temporary dir to ISyncTarget
Instead of passing the temporary directory to the ISyncTarget and
relay on proper behaviour of its implementation, a downloader
is passed to BaseSyncAsset. This downloader uses a well known file
name and downloads to the temporary directory. So everything is in
control of the SyncJob.

The ISyncTarget only needs to consume the downloaded file - if
there is any available.