refact: introduce job execution queue
The sync job is NOT allowed to run twice for the same tenant.
So far it has used a global static semaphore to ensure no
other job runs at the same time.
The global semaphore is bad, as it will block even if the job
runs with a different tenant settings.

Now an execution queue takes care of that. Since it is used
as a singleton within the dependency injection scope but each
tenant uses its own scope, no tenant will block other tenants.