feature:(auth) provide in-memory storage of token data
A lot of sync target processes are long running processes
kept in memory. So the Smint.io authentication token data
can be kept in memory and does not need to be made persistent
in a database or filesystem.

As this is commonly the case with most of sync targets integration,
it is good to have this class as part of the base library.