feature:(data) provide generic JSON based storage for data
JSON is a very common serialization format. As it is widely
used its usage is easy and thus can be used to store data to
and load data from a file.

A generic class is introduced, as this is a very common case for
most of the integration sync targets.