refact:(auth) make authenticator provide its auth data storage
When using multiple authenticators because multiple tenants
or target systems are in use, it can be cumbersome and
error prune to tweak dependency injection that both - consumer and
producer of authentication data - use the same storage.
Hence the authenticator provides the storage it uses to the
consumer. Then it is guaranteed, that both use the same
auth data storage!