refact:(auth) provider interfaces for OAuth and JWT authenticator
By using these interfaces, it is clear to the authentication
consumers and the dependency injection, what kind of data these
authenticators are using to store their authentication data.

With the help of these interfaces, the dependency injection system
is able to strongly type and test the authentication data that
is produced by these authenticators.