refact:(auth) make "SmintIoSystemBrowserAuthenticatorImpl" inherit from "SmintIoAuthenticatorImpl"
In order to use the extended browser based authenticator as
a replacement for the simpler headless refresher, the browser
based authenticator must be inherit from the latter.

Initially this was impossible as both classes had different,
distinct inheritcance trees (base classes). So, to make it
work-out well, both classes are now facades to the real
implementing instances - OAuth in this case.