feature:(hostedService) introduce "SingleTargetSyncService" as a hosted service wrapper
To enable execution of multiple, different sync clients, these must be
separated into different dependency injection spaces. Since the
timer and pusher hosted services directly depend on the target
implementation, these can not be in common space. If not in
common space, they can not be started automatically.

In order to utilize the same setup when running as a stand-alone
client, this new wrapper is used. It will initiate the
DI space for the client and run the timer and pusher service.