feature:(DI) provide separate target DI scope with "Autofac"
To run various synchronizers to different targets
simultaneously, these must be separated with different
DI scopes. Autofac is used to achieve this and special
target factories are used to set-up these separate
scopes. The target factories may live in global
scope and are used to create a target sync client inside
its separate, child scope.

More changes are needed to really make that work, but this
is the basis to run multiple sync clients to different
targets in parallel.